Refactor mutliple references
